Move Loader to icons dir
