+++ frontend/src/lib/components/icons/Loader.svelte
