+++ frontend/src/routes/rooms/[roomId]/HostControls.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import TgButton from "$lib/components/base/TgButton.svelte";
   import TgParagraph from "$lib/components/base/TgParagraph.svelte";
-  import Loader from "$lib/components/Loader.svelte";
+  import Loader from "$lib/components/icons/Loader.svelte";
   import type { Room } from "$lib/services/rooms";
   import { createEventDispatcher } from "svelte";
